requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
email-validator
python-dotenv
aiofiles
httpx